third_party/blink/web_tests/platform/generic/fast/forms/time-multiple-fields/time-multiple-fields-focus-expected.txt

Check if focus() for focused input does not change focused sub-field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Move focus to the second sub-field:
PASS document.activeElement is timeInput
PASS shadowPseudoIdOfFocused(timeInput) is "-webkit-datetime-edit-minute-field"
PASS timeInput.focus(); shadowPseudoIdOfFocused(timeInput) is "-webkit-datetime-edit-minute-field"
Click on a delimiter between sub-fields, then check if focused element is not changed:
PASS shadowPseudoIdOfFocused(timeInput) is "-webkit-datetime-edit-minute-field"
PASS successfullyParsed is true

TEST COMPLETE
